+++ lib/runners/processor/eslint.rb
@@ -25,7 +25,7 @@ module Runners
       )
     end
 
-    register_config_schema(name: :eslint, schema: SCHEMA.config)
+    register_config_schema SCHEMA.config
 
     CONSTRAINTS = {
       "eslint" => Gem::Requirement.new(">= 5.0.0", "< 8.0.0").freeze,
@@ -55,7 +55,7 @@ module Runners
     end
 
     def setup
-      add_warning_for_deprecated_option :dir, to: :target
+      warnings.add_warning_for_deprecated_option(config: config, analyzer: analyzer_id, old: :dir, new: :target)
 
       begin
         install_nodejs_deps constraints: CONSTRAINTS