+++ lib/runners/processor/haml_lint.rb
@@ -22,7 +22,7 @@ module Runners
       )
     end
 
-    register_config_schema(name: :haml_lint, schema: SCHEMA.config)
+    register_config_schema SCHEMA.config
 
     GEM_NAME = "haml_lint".freeze
     REQUIRED_GEM_NAMES = ["rubocop"].freeze
@@ -53,7 +53,7 @@ module Runners
     end
 
     def setup
-      add_warning_for_deprecated_option :file, to: :target
+      warnings.add_warning_for_deprecated_option(config: config, analyzer: analyzer_id, old: :file, new: :target)
 
       setup_haml_lint_config