+++ test/smokes/eslint/expectations.rb
@@ -1,6 +1,6 @@
 s = Runners::Testing::Smoke
 
-default_version = "7.22.0"
+default_version = "7.23.0"
 
 s.add_test(
   "no_config",
@@ -430,7 +430,7 @@ s.add_test(
       id: "@typescript-eslint/no-unused-vars",
       message: "'x' is assigned a value but never used.",
       links: %w[
-        https://github.com/typescript-eslint/typescript-eslint/blob/v4.18.0/packages/eslint-plugin/docs/rules/no-unused-vars.md
+        https://github.com/typescript-eslint/typescript-eslint/blob/v4.20.0/packages/eslint-plugin/docs/rules/no-unused-vars.md
       ],
       path: "index.ts",
       location: { start_line: 1, start_column: 7, end_line: 1, end_column: 8 },
@@ -540,7 +540,7 @@ s.add_test(
   "package_lock_without_package_json",
   type: "success",
   issues: [],
-  analyzer: { name: "ESLint", version: "7.22.0" }
+  analyzer: { name: "ESLint", version: default_version }
 )
 
 s.add_test(