+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Updated environments:
 
 Updated tools:
 
+- **Cppcheck** 2.3 -> 2.4.1 [#2229](https://github.com/sider/runners/pull/2229)
 - **ktlint** 0.39.0 -> 0.41.0 [#2205](https://github.com/sider/runners/pull/2205)
 
 Misc: