+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Updated environments:
 
 Updated tools:
 
+- **Checkstyle** 8.41 -> 8.41.1 [#2247](https://github.com/sider/runners/pull/2247)
 - **CoffeeLint** 4.1.3 -> 4.1.4 [#2248](https://github.com/sider/runners/pull/2248)
 - **Cppcheck** 2.3 -> 2.4.1 [#2229](https://github.com/sider/runners/pull/2229)
 - **GolangCI-Lint** 1.38.0 -> 1.39.0 [#2239](https://github.com/sider/runners/pull/2239)