+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Updated tools:
 - **ESLint** 7.22.0 -> 7.23.0 [#2241](https://github.com/sider/runners/pull/2241)
 - **GolangCI-Lint** 1.38.0 -> 1.39.0 [#2239](https://github.com/sider/runners/pull/2239)
 - **ktlint** 0.39.0 -> 0.41.0 [#2205](https://github.com/sider/runners/pull/2205)
+- **LanguageTool** 5.2 -> 5.3 [#2251](https://github.com/sider/runners/pull/2251)
 - **PMD CPD** 6.32.0 -> 6.33.0 [#2246](https://github.com/sider/runners/pull/2246)
 - **PMD Java** 6.32.0 -> 6.33.0 [#2244](https://github.com/sider/runners/pull/2244)
 - **RuboCop** 1.11.0 -> 1.12.0 [#2240](https://github.com/sider/runners/pull/2240)